+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+### Added
+- Added `contour` function
+
 ## 0.6.1 - 2021-07-03
 ### Fixed
 - Bug in `gdstk::read_oas` and `Library.read_oas`